+++ puree/src/androidTest/java/com/cookpad/puree/async/AsyncResultTest.java
@@ -1,26 +1,27 @@
 package com.cookpad.puree.async;
 
-import android.support.test.runner.AndroidJUnit4;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertThat;
+import android.support.test.runner.AndroidJUnit4;
+
+import static org.hamcrest.Matchers.*;
+import static org.junit.Assert.*;
 
 @RunWith(AndroidJUnit4.class)
 public class AsyncResultTest {
+
     @Test
-    public void get() throws InterruptedException {
-        {
-            AsyncResult asyncResult = new AsyncResult();
-            asyncResult.success();
-            assertThat(asyncResult.get(), is(true));
-        }
-        {
-            AsyncResult asyncResult = new AsyncResult();
-            asyncResult.fail();
-            assertThat(asyncResult.get(), is(false));
-        }
+    public void getWithSuccess() throws Exception {
+        AsyncResult asyncResult = new AsyncResult();
+        asyncResult.success();
+        assertThat(asyncResult.get(), is(true));
+    }
+
+    @Test
+    public void getWithFail() throws Exception {
+        AsyncResult asyncResult = new AsyncResult();
+        asyncResult.fail();
+        assertThat(asyncResult.get(), is(false));
     }
 }
