+++ puree/src/androidTest/java/com/cookpad/puree/retryable/RetryableTaskRunnerTest.java
@@ -1,22 +1,22 @@
-package com.cookpad.puree;
-
-import android.os.Handler;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
-
-import com.cookpad.puree.retryable.RetryableTaskRunner;
+package com.cookpad.puree.retryable;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import android.os.Handler;
+import android.support.test.InstrumentationRegistry;
+import android.support.test.runner.AndroidJUnit4;
+
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
-import static org.junit.Assert.fail;
+import static org.hamcrest.Matchers.*;
+import static org.junit.Assert.*;
 
 @RunWith(AndroidJUnit4.class)
 public class RetryableTaskRunnerTest {
+
     private Handler handler;
 
     @Before
@@ -30,22 +30,18 @@ public class RetryableTaskRunnerTest {
     }
 
     @Test
-    public void ensureToCallMeAfterSetTime() {
+    public void ensureToCallMeAfterSetTime() throws Exception {
         final CountDownLatch latch = new CountDownLatch(1);
 
-        RetryableTaskRunner retryableTaskRunner = new RetryableTaskRunner(new Runnable() {
+        RetryableTaskRunner task = new RetryableTaskRunner(new Runnable() {
             @Override
             public void run() {
                 latch.countDown();
             }
         }, 10, 5, handler);
 
-        retryableTaskRunner.tryToStart();
+        task.tryToStart();
 
-        try {
-            latch.await(30, TimeUnit.MILLISECONDS);
-        } catch (InterruptedException e) {
-            fail(e.getMessage());
-        }
+        assertThat(latch.await(30, TimeUnit.MILLISECONDS), is(true));
     }
 }
